+++ CHANGELOG.md
@@ -14,6 +14,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.0.7] - 2025-11-02
+### Changed
+- Project UI updated to be more intuitive.
+
+### Added
+- Copy as HTML on Generated Blog Posts.
+- PDF Generation for blog posts.
+- Centralized location of all AI models used in the app.
+- Referrer model to display banners for expected referrer like producthunt
+
+
 ## [0.0.7] - 2025-11-01
 ### Changed
 - update to user-settings page